+++ src/main/java/com/verisign/storm/metrics/GraphiteMetricsConsumer.java
@@ -18,13 +18,12 @@ import backtype.storm.metric.api.IMetricsConsumer;
 import backtype.storm.task.IErrorReporter;
 import backtype.storm.task.TopologyContext;
 import com.google.common.base.Throwables;
-import com.verisign.storm.metrics.graphite.ConnectionFailureException;
 import com.verisign.storm.metrics.reporters.AbstractReporter;
+import com.verisign.storm.metrics.util.ConnectionFailureException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.Collection;
 import java.util.HashMap;
@@ -36,10 +35,11 @@ import java.util.Map;
  * To use, add this to your topology's configuration:
  *
  * To report to a Graphite endpoint:
+ *
  * <pre>
  * {@code
  *   conf.registerMetricsConsumer(com.verisign.storm.metrics.GraphiteMetricsConsumer.class, 1);
- *   conf.put("metrics.reporter.name", "com.verisign.storm.metrics.reporters.GraphiteReporter");
+ *   conf.put("metrics.reporter.name", "com.verisign.storm.metrics.reporters.graphite.GraphiteReporter");
  *   conf.put("metrics.graphite.host", "<GRAPHITE HOSTNAME>");
  *   conf.put("metrics.graphite.port", "<GRAPHITE PORT>");
  *   conf.put("metrics.graphite.prefix", "<DOT DELIMITED PREFIX>");
@@ -48,16 +48,28 @@ import java.util.Map;
  * }
  * </pre>
  *
- * To report to a Kafka endpoint:
+ * To report Avro-encoded metrics to a Kafka endpoint:
  * <pre>
  * {@code
  *   conf.registerMetricsConsumer(com.verisign.storm.metrics.GraphiteMetricsConsumer.class, 1);
- *   conf.put("metrics.reporter.name", "com.verisign.storm.metrics.reporters.KafkaReporter");
+ *   conf.put("metrics.reporter.name", "com.verisign.storm.metrics.reporters.kafka.AvroKafkaReporter");
  *   conf.put("metrics.graphite.prefix", "storm.cluster.metrics");
  *   conf.put("metrics.kafka.topic", "graphingMetrics");
  *   conf.put("metrics.kafka.metadata.broker.list", "kafka1.example.com:9092,kafka2.example.com:9092");
  * }
- * </pre>  
+ * </pre>
+ *
+ * To report Avro-encoded metrics to a Confluent Schema-Registry enabled Kafka endpoint:
+ * <pre>
+ * {@code
+ *   conf.registerMetricsConsumer(com.verisign.storm.metrics.GraphiteMetricsConsumer.class, 1);
+ *   conf.put("metrics.reporter.name", "com.verisign.storm.metrics.reporters.kafka.SchemaRegistryKafkaReporter");
+ *   conf.put("metrics.graphite.prefix", "storm.cluster.metrics");
+ *   conf.put("metrics.kafka.topic", "graphingMetrics");
+ *   conf.put("metrics.kafka.metadata.broker.list", "kafka1.example.com:9092,kafka2.example.com:9092");
+ *   conf.put("metrics.kafka.schema.registry.url", "http://schemaregistry.example.com:8081");
+ * }
+ * </pre>   
  *
  *
  * Or edit the storm.yaml config file:
@@ -69,7 +81,7 @@ import java.util.Map;
  *     - class: "com.verisign.storm.metrics.GraphiteMetricsConsumer"
  *       parallelism.hint: 1
  *       argument:
- *         metrics.reporter.name: "com.verisign.storm.metrics.reporters.GraphiteReporter"
+ *         metrics.reporter.name: "com.verisign.storm.metrics.reporters.graphite.GraphiteReporter"
  *         metrics.graphite.host: "<GRAPHITE HOSTNAME>"
  *         metrics.graphite.port: "<GRAPHITE PORT>"
  *         metrics.graphite.prefix: "<DOT DELIMITED PREFIX>"
@@ -77,21 +89,34 @@ import java.util.Map;
  * }
  * </pre>
  *
- * To report to a Kafka endpoint:
+ * To report Avro-encoded metrics to a Kafka endpoint:
  * <pre>
  * {@code
  *  topology.metrics.consumer.register:
  *    - class: "com.verisign.storm.metrics.GraphiteMetricsConsumer"
  *      parallelism.hint: 1
  *      argument:      
- *        metrics.reporter.name: "com.verisign.storm.metrics.reporters.KafkaReporter"
+ *        metrics.reporter.name: "com.verisign.storm.metrics.reporters.kafka.AvroKafkaReporter"
  *        metrics.graphite.prefix: "storm.cluster.metrics"
  *        metrics.kafka.topic: "graphingMetrics"
  *        metrics.kafka.metadata.broker.list: "kafka1.example.com:9092,kafka2.example.com:9092"
  * }
  * </pre> 
  *
-
+ * To report Avro-encoded metrics to a Confluent Schema-Registry enabled Kafka endpoint:
+ * <pre>
+ * {@code
+ *  topology.metrics.consumer.register:
+ *    - class: "com.verisign.storm.metrics.GraphiteMetricsConsumer"
+ *      parallelism.hint: 1
+ *      argument:
+ *        metrics.reporter.name: "com.verisign.storm.metrics.reporters.kafka.SchemaRegistryKafkaReporter"
+ *        metrics.graphite.prefix: "storm.cluster.metrics"
+ *        metrics.kafka.topic: "graphingMetrics"
+ *        metrics.kafka.metadata.broker.list: "kafka1.example.com:9092,kafka2.example.com:9092"
+ *        metrics.kafka.schema.registry.url: "http://schemaregistry.example.com:8081"
+ * }
+ * </pre>
  */
 public class GraphiteMetricsConsumer implements IMetricsConsumer {
 
@@ -104,7 +129,7 @@ public class GraphiteMetricsConsumer implements IMetricsConsumer {
   private String stormId;
   private Map reporterConfig;
 
-  protected AbstractReporter adapter;
+  protected AbstractReporter reporter;
 
   protected String getStormId() {
     return stormId;
@@ -117,11 +142,23 @@ public class GraphiteMetricsConsumer implements IMetricsConsumer {
   @Override
   public void prepare(Map config, Object registrationArgument, TopologyContext context, IErrorReporter errorReporter) {
 
-    Map configMap = new HashMap();
-    configMap.putAll(config);
-    if (registrationArgument instanceof Map) {
+    Map configMap = new HashMap<Object, Object>();
+    if (config != null) {
+      LOG.info("Configuration parameter: {}", config.toString());
+      configMap.putAll(config);
+    }
+    else {
+      LOG.warn("No reference to configuration parameter config found.");
+    }
+
+    if (registrationArgument != null && registrationArgument instanceof Map) {
+      LOG.info("Registration argument: {}", registrationArgument.toString());
       configMap.putAll((Map) registrationArgument);
     }
+    else {
+      LOG.warn("No reference to configuration parameter registrationArgument found");
+    }
+
     reporterConfig = configMap;
 
     if (reporterConfig.containsKey(GRAPHITE_PREFIX_OPTION)) {
@@ -132,23 +169,23 @@ public class GraphiteMetricsConsumer implements IMetricsConsumer {
     }
 
     try {
-      adapter = configureAdapter(reporterConfig);
+      reporter = configureReporter(reporterConfig);
     }
     catch (Exception e) {
-      LOG.error("Error configuring metrics adapter:" + e.getMessage());
+      LOG.error("Error configuring metrics reporter:" + e.getMessage());
       errorReporter.reportError(e);
     }
 
     stormId = context.getStormId();
   }
 
-  private AbstractReporter configureAdapter(@SuppressWarnings("rawtypes") Map reporterConfig)
+  private AbstractReporter configureReporter(@SuppressWarnings("rawtypes") Map reporterConfig)
       throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException,
       InvocationTargetException {
     String className = (String) reporterConfig.get(REPORTER_NAME);
     Class reporterClass = Class.forName(className);
-    Constructor ctor = reporterClass.getConstructor(Map.class);
-    AbstractReporter reporter = (AbstractReporter) ctor.newInstance(reporterConfig);
+    AbstractReporter reporter = (AbstractReporter) reporterClass.newInstance();
+    reporter.prepare(reporterConfig);
 
     return reporter;
   }
@@ -169,29 +206,27 @@ public class GraphiteMetricsConsumer implements IMetricsConsumer {
         continue;
       }
 
-      // Most data points contain a Map as a value.
-      if (dataPoint.value instanceof Map) {
+      if (dataPoint.value == null) {
+        LOG.warn("Data point with name {} has a value of null. Discarding data point", dataPoint.name);
+      }
+      else if (dataPoint.value instanceof Map) {
         Map<String, Object> dataMap = (Map<String, Object>) dataPoint.value;
         Map<String, Double> bufferMap = new HashMap<String, Double>();
 
-        for (String key : dataMap.keySet()) {
-          if (dataMap.get(key) instanceof Number) {
-            bufferMap.put(key, ((Number) dataMap.get(key)).doubleValue());
-          }
-          else {
-            LOG.warn("Unrecognized metric value of type {} received. Discarding metric.",
-                dataMap.get(key).getClass().getName());
+        for (Map.Entry<String, Object> entry : dataMap.entrySet()) {
+          Double dblValue = convertToDoubleValue(entry.getKey(), entry.getValue());
+          if (dblValue != null) {
+            bufferMap.put(entry.getKey(), dblValue);
           }
         }
 
-        appendToBuffer(metricPrefix, bufferMap, taskInfo.timestamp);
+        appendToReporterBuffer(metricPrefix, bufferMap, taskInfo.timestamp);
       }
-
       else if (dataPoint.value instanceof Number) {
         Double dblValue = ((Number) dataPoint.value).doubleValue();
         HashMap<String, Double> metric = new HashMap<String, Double>();
         metric.put(dataPoint.name + "value", dblValue);
-        appendToBuffer(metricPrefix, metric, taskInfo.timestamp);
+        appendToReporterBuffer(metricPrefix, metric, taskInfo.timestamp);
       }
       else {
         LOG.warn("Unrecognized metric value of type {} received. Discarding metric.",
@@ -202,6 +237,33 @@ public class GraphiteMetricsConsumer implements IMetricsConsumer {
     graphiteDisconnect();
   }
 
+  private Double convertToDoubleValue(String key, Object value) {
+    if (value != null) {
+      Double result;
+
+      if (value instanceof String) {
+        try {
+          result = Double.parseDouble((String) value);
+        }
+        catch (NumberFormatException e) {
+          LOG.warn("Failed to parse metric with key {} and value of {} into a Double. Discarding metric.", key, value);
+          result = null;
+        }
+      }
+      else if (value instanceof Number) {
+        result = ((Number) value).doubleValue();
+      }
+      else {
+        result = null;
+      }
+
+      return result;
+    }
+    else {
+      LOG.warn("Metric with key {} has a value of null. Discarding metric.", key);
+      return null;
+    }
+  }
   /**
    * Constructs a fully qualified metric prefix.
    *
@@ -231,46 +293,47 @@ public class GraphiteMetricsConsumer implements IMetricsConsumer {
 
   protected void graphiteConnect() {
     try {
-      adapter.connect();
+      reporter.connect();
     }
     catch (ConnectionFailureException e) {
       String trace = Throwables.getStackTraceAsString(e);
-      LOG.error("Could not connect to adapter backend " + adapter.getBackendFingerprint() + ": " + trace);
+      LOG.error("Could not connect to reporter backend " + reporter.getBackendFingerprint() + ": " + trace);
     }
   }
 
-  protected void appendToBuffer(String prefix, Map<String, Double> metrics, long timestamp) {
-    if (adapter != null) {
-      adapter.appendToBuffer(prefix, metrics, timestamp);
+  protected void appendToReporterBuffer(String prefix, Map<String, Double> metrics, long timestamp) {
+    if (reporter != null) {
+      reporter.appendToBuffer(prefix, metrics, timestamp);
     }
   }
 
   protected void emptyBuffer() {
-    adapter.emptyBuffer();
+    reporter.emptyBuffer();
   }
 
   protected void sendMetrics() {
     try {
-      if (adapter != null) {
-        adapter.sendBufferContents();
+      if (reporter != null) {
+        reporter.sendBufferContents();
       }
     }
     catch (IOException e) {
       String trace = Throwables.getStackTraceAsString(e);
-      String msg = "Could not send metrics update to backend server " + adapter.getBackendFingerprint() + ": " + trace +
-          " (" + adapter.getFailures() + " failed attempts so far)";
+      String msg = "Could not send metrics update to backend server " + reporter.getBackendFingerprint() + ": " + trace
+          +
+          " (" + reporter.getFailures() + " failed attempts so far)";
       LOG.error(msg);
     }
   }
 
   protected void graphiteDisconnect() {
-    if (adapter != null) {
+    if (reporter != null) {
       try {
-        adapter.disconnect();
+        reporter.disconnect();
       }
       catch (ConnectionFailureException e) {
         String trace = Throwables.getStackTraceAsString(e);
-        LOG.error("Could not disconnect from adapter backend " + adapter.getBackendFingerprint() + ": " + trace);
+        LOG.error("Could not disconnect from reporter backend " + reporter.getBackendFingerprint() + ": " + trace);
       }
     }
   }