+++ src/main/java/com/verisign/storm/metrics/reporters/AbstractReporter.java
@@ -1,15 +1,17 @@
 package com.verisign.storm.metrics.reporters;
 
-import com.verisign.storm.metrics.graphite.ConnectionFailureException;
+import com.verisign.storm.metrics.util.ConnectionFailureException;
 
 import java.io.IOException;
 import java.util.Map;
 
 public abstract class AbstractReporter {
 
-  public AbstractReporter(Map<String, Object> conf) {
+  public AbstractReporter() {
   }
 
+  public abstract void prepare(Map<String, Object> conf);
+  
   public abstract void connect() throws ConnectionFailureException;
 
   public abstract void disconnect() throws ConnectionFailureException;